lambdas/transformation_lambda/pyproject.toml
[project]
name = "mbtp_splunk_cloudwatch_transformation"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "Thomas Nicholson", email = "thomas.nicholson@accenture.com"},
]
dependencies = [
    "boto3>=1.35.66",
    "cerberus>=1.3.7",
    "ruamel-yaml>=0.18.10",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "n/a"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
test = "pdm run pytest -rP --junitxml=reports/pytest.xml --cov=src --cov-report=xml:reports/coverage.xml --cov-report term-missing tests/"
lint = "pdm run pylint src/mbtp_splunk_cloudwatch_transformation/"
format = "pdm run black --check src/mbtp_splunk_cloudwatch_transformation/"
check = {composite = ["format", "lint", "test"]}
requirements = "pdm export -o requirements.txt --without-hashes --prod"

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.10.0",
    "pytest>=8.3.3",
    "pylint>=3.3.1",
    "pytest-cov>=6.0.0",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.0",
]

[tool.pylint.'MESSAGES CONTROL']
disable=["logging-fstring-interpolation","logging-not-lazy"]

[tool.pytest.ini_options]
env = [
    "AWS_ACCESS_KEY_ID=testing",
    "AWS_SECRET_ACCESS_KEY=testing",
    "AWS_SECURITY_TOKEN=testing",
    "AWS_SESSION_TOKEN=testing",
    "AWS_REGION=eu-west-1",
    "CONFIG_S3_BUCKET=CONFIG_S3_BUCKET",
    "CONFIG_S3_KEY=CONFIG_S3_KEY",
]
